Revert "exfat: push 5.6-rc1 kernel"
This reverts commit 5c8a1d76a57ba62b446dafb7e661349ec78183bc.

+++ drivers/usb/storage/transport.c
@@ -1284,7 +1284,8 @@ int usb_stor_Bulk_transport(struct scsi_cmnd *srb, struct us_data *us)
 
 		} else {
 			residue = min(residue, transfer_length);
-			scsi_set_resid(srb, max(scsi_get_resid(srb), residue));
+			scsi_set_resid(srb, max(scsi_get_resid(srb),
+			                                       (int) residue));
 		}
 	}
